fix some pir case in windows
