+++ test/deprecated/legacy_test/CMakeLists.txt
@@ -657,8 +657,9 @@ endif()
 set_tests_properties(test_inplace_addto_strategy_deprecated PROPERTIES TIMEOUT
                                                                        120)
 
-set(TEST_CINN_OPS test_slice_op_deprecated test_layer_norm_op_deprecated
-                  test_where_op test_instance_norm_op_deprecated)
+set(TEST_CINN_OPS
+    test_slice_op_deprecated test_layer_norm_op_deprecated test_where_op
+    test_instance_norm_op_deprecated test_group_norm_op_deprecated)
 
 foreach(TEST_CINN_OP ${TEST_CINN_OPS})
   if(WITH_CINN)
@@ -738,3 +739,4 @@ set_tests_properties(test_conv3d_layer_deprecated PROPERTIES TIMEOUT 100)
 set_tests_properties(test_attribute_var_deprecated PROPERTIES TIMEOUT 100)
 set_tests_properties(test_inference_api_deprecated PROPERTIES TIMEOUT 100)
 set_tests_properties(test_lbfgs_deprecated PROPERTIES TIMEOUT 100)
+set_tests_properties(test_group_norm_op_deprecated PROPERTIES TIMEOUT 1000)
