+++ test/legacy_test/CMakeLists.txt
@@ -886,7 +886,6 @@ set_tests_properties(
   test_buffer_shared_memory_reuse_pass_and_fuse_optimization_op_pass
   PROPERTIES TIMEOUT 120)
 set_tests_properties(test_pool_max_op PROPERTIES TIMEOUT 500)
-set_tests_properties(test_group_norm_op PROPERTIES TIMEOUT 1000)
 set_tests_properties(test_imperative_transformer_sorted_gradient
                      PROPERTIES TIMEOUT 120)
 set_tests_properties(test_imperative_optimizer PROPERTIES TIMEOUT 250)
@@ -1029,7 +1028,6 @@ set(TEST_CINN_OPS
     test_batch_norm_op_prim_nchw
     test_batch_norm_op_prim_nhwc
     test_dropout_op
-    test_group_norm_op
     test_tile_op
     test_sum_op
     test_squeeze2_op
